# openapi/openapi.yml
openapi: 3.0.1
info:
  title: Meddle
  description: This is meddle, an app that keeps track of medications with notification
    feature.
  contact:
    email: info@decagonhq.com
  version: 3.0.0
servers:
  - url: http://ec2-3-143-47-141.us-east-2.compute.amazonaws.com/api/v1
  - url: https://ec2-3-143-47-141.us-east-2.compute.amazonaws.com/api/v1
  - url: http://localhost:8081/api/v1
paths:
...
  /auth/login:
    post:
      tags:
        - user
      summary: Logs user into the system
      operationId: loginUser
      requestBody:
        description: login user to the system
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        200:
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        400:
          description: invalid email/password supplied
          content: {}
        401:
          description: inactive user or wrong password
          content: { }
        422:
          description: email does not exist, system does not recognise email
          content: { }
        500:
          description: internal server error
          content: { }
      x-codegen-request-body-name: user
...
  /password/forgot:
    post:
      tags:
        - user
      summary: Sends the user email when forgot password is hit
      operationId: ForgotPassword
      requestBody:
        description: Sends the user email when forgot password is hit
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
        required: true
      responses:
        200:
          description: link to reset password successfully sent
          headers:
            X-Rate-Limit:
              schema:
                type: string
                format: string
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        400:
          description: email was not sent
          content: {}
      x-codegen-request-body-name: ForgortPassword
...
components:
  schemas:
    UserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
    ForgotPasswordResponse:
      type: object
      properties:
        message:
          type: string
          format: object
...
    LoginResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/loginResponseData'
        errors:
          type: string
          example: ""
        message:
          type: string
          example: user created successfully
        status:
          type: integer
          example: 200
...
    loginResponseData:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Ken
        phone_number:
          type: string
          example: "+2349043556XXX"
        email:
          type: string
          example: ken@gmail.com
        accessToken:
          type: string
          example: Rbhfwi2PUXndOWVlUpsy0
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes